@@ -1,5 +1,5 @@
 module github.com/kinbiko/bugsnag
 
-go 1.16
+go 1.20
 
-require github.com/kinbiko/jsonassert v1.1.0
+require github.com/kinbiko/jsonassert v1.1.1
