@@ -1,12 +1,18 @@
+// Package builds is a package for sending build requests to the Bugsnag Build API, as defined here:
+// https://bugsnagbuildapi.docs.apiary.io/
 package builds
 
 import (
 	"bytes"
+	"context"
 	"encoding/json"
 	"fmt"
 	"io"
 	"net/http"
+	"time"
 )
+
+const timeout = 10 * time.Second
 
 // DefaultPublisher returns a publisher for requests against the Bugsnag SaaS
 // endpoint. If you need to target your own on-premise installation, please use
@@ -23,7 +29,7 @@ func NewPublisher(endpoint string) *Publisher {
 }
 
 // Publisher is a type for sending build requests to the Bugsnag Build API, as defined here:
-//https://bugsnagbuildapi.docs.apiary.io/
+// https://bugsnagbuildapi.docs.apiary.io/
 type Publisher struct {
 	endpoint string
 }
@@ -34,13 +40,14 @@ func (p *Publisher) Publish(req *JSONBuildRequest) error {
 		return fmt.Errorf("publisher created incorrectly; please use NewPublisher or DefaultPublisher to construct your builds.Publisher")
 	}
 
-	b, err := json.Marshal(req)
+	jsonBody, err := json.Marshal(req)
 	if err != nil {
 		return fmt.Errorf("unable to marshal JSON: %w", err)
 	}
 
-	httpReq, err := http.NewRequest("POST", p.endpoint, bytes.NewBuffer(b))
-
+	ctx, cancel := context.WithTimeout(context.Background(), timeout)
+	defer cancel()
+	httpReq, err := http.NewRequestWithContext(ctx, http.MethodPost, p.endpoint, bytes.NewBuffer(jsonBody))
 	if err != nil {
 		return fmt.Errorf("error when POST-ing request to '%s': %w", p.endpoint, err)
 	}
@@ -48,12 +55,15 @@ func (p *Publisher) Publish(req *JSONBuildRequest) error {
 
 	httpRes, err := http.DefaultClient.Do(httpReq)
 	if err != nil {
-		return err
+		return fmt.Errorf("unable to make HTTP call: %w", err)
 	}
+	defer func() {
+		_ = httpRes.Body.Close()
+	}()
 
 	got, err := io.ReadAll(httpRes.Body)
 	if err != nil {
-		return err
+		return fmt.Errorf("unable to read response body: %w", err)
 	}
 
 	type response struct {
@@ -63,7 +73,7 @@ func (p *Publisher) Publish(req *JSONBuildRequest) error {
 	}
 	var res response
 	if err := json.Unmarshal(got, &res); err != nil {
-		return err
+		return fmt.Errorf("unable to read response body as JSON: %w", err)
 	}
 	if res.Status == "error" {
 		return fmt.Errorf("error when sending message: %s", res.Errors[0])
