@@ -1,14 +1,21 @@
 package main
 
-import "strings"
+import (
+	"log"
+	"strings"
+)
 
 func splitByEquals(strs []string) map[string]string {
-	m := map[string]string{}
+	kvps := map[string]string{}
 	for _, kvp := range strs {
 		if kvp != "" {
 			pair := strings.SplitN(kvp, "=", 2)
-			m[pair[0]] = pair[1]
+			kvps[pair[0]] = pair[1]
 		}
 	}
-	return m
+	return kvps
+}
+
+func logf(msg string, args ...interface{}) {
+	log.Printf(msg, args...)
 }
